Refactor: create file for each handler
